+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add cleaner for Application Port Profiles.
 - Add `compatibleProviders` to `Chart.yaml`.
 - Add use of runtime/default seccomp profile.
+- Add cleaner for NamedDisks(Volumes).
 
 ## [0.2.0] - 2022-11-24